c file

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -6,8 +6,8 @@
  */
 int main(void)
 {
-	unsigned long int number = 612852475143;
-	unsigned long int count = 2, count2, max = 612852475143;
+	long long number = 612852475143;
+	long long count = 2, count2, max = 612852475143;
 	int status;
 
 	while (count < number)
